comfyui/requirements.txt
torchaudio
torchlibrosa
torchvision
diffusers
accelerate
datasets
librosa
wandb
tqdm